Update patch set 5
Patch Set 5: Verified+1 Code-Review+2

We'll revisit the app install code at a later date since it's going to need a lot of testing.

We ran into a bunch of bugs the last time which is why all the hacks, now that we know what to look for we can take our time to do it properly.

Patch-set: 5
Reviewer: Gerrit User 1000000 <1000000@35266860-6252-469c-a67d-185064479fc0>
Label: Code-Review=+2
Label: Verified=+1