Send MicroG Check-in broadcast after provisioning is complete
This is required because we prevent from MicroG from making a check-in while not provisioned to ensure that it doesn't make check-in connections to Google when the user has not yet enabled it.

Even without the other MicroG fixes submitted to upstream, this might already help with the issue of apps thinking that push message support is not available after first install.

Change-Id: Iefbb8b0f3fcabe940114f33afd78c34d9cb5fce7