Switch network setup and missing SIM activity order
Missing SIM activity offers to setup an eSIM which can only be done if
connected to a network. Therefore, show network setup activity first
and, if a SIM is still missing after it, show the missing SIM activity.
If connected to a network, show the option to setup an eSIM. Otherwise,
hide it.

Change-Id: Ieaa51940d916489212a35ee95cc81a33c6cabe9a