Disable PartnerReceiver
PartnerReceiver only serves as a marker; it does not need to be enabled.
Disabling it removes an unnecessary entry point into the Setup Wizard.

Broadcasting com.android.setupwizard.action.PARTNER_CUSTOMIZATION
would otherwise result in SetupWizardApp's onCreate being unnecessarily
executed, which has posed problems in the past. That should be less
of a concern today due to an early return if SUW is complete, so this
change is merely an additional measure.

Having a PARTNER_CUSTOMIZATION receiver indicates to the setupdesign
library that we may provide customized resources for other components
that utilize the library. It does not need or expect to ever actually
handle a broadcast. The library also matches disabled components in
its search for matching receivers.

Test: Manual: Run `adb shell setprop log.tag.SetupWizardApp VERBOSE`.
Run `adb shell am broadcast -c android.intent.category.DEFAULT \
-a com.android.setupwizard.action.PARTNER_CUSTOMIZATION \
org.lineageos.setupwizard`. A logcat should not contain
`SetupWizardApp: onCreate()`.

Issue: calyxos#1740
Change-Id: Idf78361ed415e403fcce8e1c6ef5a6baddf57fb4